Move projects to home and add a contact section
