--- src/components/views/CV/CV.module.css
@@ -1,5 +1,0 @@
-.wrapper {
-    width: 100%;
-    color: var(--black);
-    text-align: center;
-}
