+++ src/components/views/Home/Home.jsx
@@ -1,22 +1,25 @@
-import { useNavigate } from "react-router-dom";
-import YellowButton from "../../Reusable/YellowButton/YellowButton";
-import WhiteButton from "../../Reusable/WhiteButton/WhiteButton";
+import projectData from "../../../assets/project-data.json";
+import ProjectCard from "../ProjectCard/ProjectCard";
+import { MdEmail } from "react-icons/md";
+import { FaPhoneAlt } from "react-icons/fa";
 import styles from "./Home.module.css";
 
 function Home() {
-    const navigate = useNavigate();
-
-    const goToProjects = () => {
-        navigate("/showcase");
-    };
-
-    const goToLinkedIn = () => {
-        window.open("https://www.linkedin.com/in/chloesapage/");
-    };
-
-    const goToGitHub = () => {
-        window.open("https://github.com/ChloeSAPage");
-    };
+    const projects = projectData.map((item, index) => {
+        return (
+            <ProjectCard
+                index={index}
+                key={item.projectName}
+                name={item.projectName}
+                description={item.description}
+                githubLink={item.githubLink}
+                link={item.link}
+                image={item.image}
+                languages={item.languages}
+                libraries={item.libraries}
+            />
+        );
+    });
 
     return (
         <div className={styles.wrapper}>
@@ -26,13 +29,17 @@ function Home() {
                 I'm a Junior Developer, with a background in microbiology. I
                 like making things.
             </p>
-            <div className={styles.buttonContainer}>
-                <YellowButton onClick={() => goToProjects()} text="Projects" />
-                <WhiteButton onClick={() => goToLinkedIn()} text="LinkedIn" />
-                <YellowButton onClick={() => goToGitHub()} text="GitHub" />
-            </div>
-
-            <h2 className={styles.aboutme}>
+            <h2 className={styles.title}>
+                C<span className={styles.underline}>ontact M</span>e
+            </h2>
+            <p>Get in touch!</p>
+            <p>
+                <MdEmail /> chloesapagepersonal@gmail.com
+            </p>
+            <p>
+                <FaPhoneAlt /> 07828124910
+            </p>
+            <h2 className={styles.title}>
                 A<span className={styles.underline}>bout M</span>e
             </h2>
             <p>
@@ -44,6 +51,11 @@ function Home() {
             <p>I like biology, specifically microbiology.</p>
             <p>I love learning new things!</p>
             <p>I like coding and problem solving</p>
+
+            <h2 className={styles.title}>
+                P<span className={styles.underline}>roject</span>s
+            </h2>
+            {projects}
         </div>
     );
 }
