--- src/components/views/Showcase/Showcase.module.css
@@ -1,11 +1,0 @@
-.wrapper {
-    color: var(--black);
-}
-
-.title {
-    text-align: center;
-}
-
-.underline {
-    border-bottom: 2px solid var(--yellow);
-}
